requirements.txt
discord.py>=2.3.0
openai>=1.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
aiohttp>=3.9.0
